feat(header): allow hiding floating headers
`headerVisible` has only effect when `headerMode` is `screen`, this effectively makes it
unintuitive to use this prop. In addition, allowing floating headers to be hidden supports a
scenario when you would like to have a splash screen, then your main screen (with a Tabs navigator)
and subsequent non-tab screens being rendered with the floating header so it transitions properly
and in iOS' native style.